Move command package to the root
