+++ command/new_pane.go
