+++ command/new_session.go
