+++ command/new_window.go
